add Lambda utils to aws-client package and refactor code to use them
