+++ packages/api/endpoints/reconciliation-reports.js
@@ -7,7 +7,7 @@ const {
   getS3Object,
   listS3ObjectsV2
 } = require('@cumulus/aws-client/S3');
-const { invoke } = require('@cumulus/ingest/aws');
+const { invoke } = require('@cumulus/aws-client/Lambda');
 
 /**
  * List all reconciliation reports
